added tests for file reading
