+++ cas-offinder-lib/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cas-offinder-lib"
+name = "cas_offinder_lib"
 version = "0.1.0"
 edition = "2021"
