+++ cas-offinder-lib/src/lib.rs
@@ -3,15 +3,16 @@
 mod bit4ops;
 mod chrom_chunk;
 mod read_fasta;
-mod read_bit2;
+mod read_2bit;
 mod read_fasta_folder;
 mod cli_err;
 
 pub use crate::bit4ops::*;
-pub use crate::read_bit2::*;
+pub use crate::read_2bit::*;
 pub use crate::read_fasta::*;
 pub use crate::read_fasta_folder::*;
 pub use crate::cli_err::*;
+pub use crate::chrom_chunk::*;
 
 pub fn addone(x:i64) -> i64{
     x+1
